move coalesce to ensembleframe API
